Refactor locations and naming
