+++ src/app/home/activity-indicators/activity-indicator.ts
@@ -1,0 +1,4 @@
+export interface ActivityIndicator {
+  label: string;
+  count: number;
+}
