+++ src/app/shared/data-service/collections.ts
@@ -1,0 +1,4 @@
+export const FIREBASE_COLLECTIONS = {
+  countrySupporters: 'country-supporters',
+  activityIndicators: 'activity-indicators'
+}
